IStorage is moved to separate file
